[devmgr] Close the service request handle if appmgr is not found
On Magenta-only builds (which do not contain 'system'),
requests to the '/svc' directory are transmitted over
clones of the "svc_root_handle" to the "svc_request_handle".
Without an appmgr, however, no one will attempt to read the
messages from the "svc_request_handle", and clients which
indefinitely block waiting for a response will be hung.

This causes, for example, "ls svc" to hang forever in Magenta.

This patch adds a short-lived "service_timeout" thread which
closes the "svc_request_handle" after a short period if the
appmgr does not consume it. This, in turn, prevents clients
from hanging.

Change-Id: I653758c358c4fa69b2442fd9f03102bf40eae7a4