[kernel] Job policies: Implement exceptions for MX_POL_BAD_HANDLE
Implement generating exceptions when a syscall's handle lookup fails
(i.e. for MX_ERR_BAD_HANDLE), gated by the MX_POL_BAD_HANDLE job
policy.

The test coverage for this covers both of the failure paths in
ProcessDispatcher::GetHandleLocked().

The size of the generated code for minipr_thread_loop() got bigger, so
we have to increase the size limit in mini-process.c.  (A cleaner
solution would be to build a separate executable for this mini
process.)

MG-599

Change-Id: I249e5735b60b0c5f2a3513112c7bc6ac0e30909a