[jemalloc] Tell jemalloc that Fuchsia overcommits
Before this change, the os_overcommits flag was false for
Fuchsia. This meant that jemalloc was attempting to decommit
pages. The code that did this had not been correctly migrated from
mmap-based code to more native magenta constructs: to "decommit"
memory it was actually remapping it with no read or write permissions.

With this change, regions are properly released. Some code paths for
Fuchsia become dead, and are removed.

Polling with the following, and waiting a few minutes for it to steady state:

echo $(( $(memgraph -vt | grep jemalloc | cut -d'{' -f2 | cut -d'}' -f1 | cut -d',' -f8 | cut -d':' -f2 | tr '\n' '+' ) 0 ))

Before:
    83959808

After:
    64643072

MG-907

Change-Id: Ida7e6ac3d5d8e0647e241bea853326f1417a9efc