[hypervisor] Ignore PIT channel 0 port out.
Linux sets the frequency of channel 0 (sending 0x12a5) and subsequently
enables TSC in x86_late_time_init (see time.c).

i8253.c sets global_clock_event to the pit and then it seems that it is
never replaced, since we don't use HPET. I'm uncertain as to whether
we actually need the PIT, or whether TSC will just be used. I could
spend a bunch more time figuring this out, but my immediate plan is
to just plow on, ignore the PIT and see where it leads us.

Change-Id: I474a3c6fafee179abc215021e00ea869bf956361