Crash device runner if ledger crashes.
If the ledger crashes, the UI still keeps running and its unclear
without looking at the log what's going on. Crashing makes it obvious
that we're in a non-recoverable state.

We could potenially retry starting the ledger but IMO the convinience is
not worth the harm.

On a side note, as a part of using namespaces/ sanboxing in framework we could
potentially be starting a separate ledger process per user in the user
runner rather than what we do now and this problem would go away.

FW-231 #done

Change-Id: I91417ef3923969d7c5b2204ea2487888f92f0e83