[uapp][launch] Update "launch" tool to use the VDSO
This tool doesn't have any test coverage, and it had bitrotted.  It
hadn't been updated to use the VDSO, and so any subprocess created
with this tool crashes.  For example:

  $ launch -s /boot/bin/launch
  [00023.025] 01118.01163> <== fatal exception: process [4397] thread [4429]
  [00023.025] 01118.01163> <== fatal page fault, PC at 0xd56a6
  ...
  Process finished with return code -1

I am resurrecting this tool so that I can add an option for setting
MX_POL_BAD_HANDLE.

Fix this issue by updating the tool to use the VDSO.  Remove the "-b"
option because that simplifies this code path and the option is
probably not useful any more.

MG-599

Change-Id: I79a8a287d75f2b54d47669faa1259451ae2d0b0b