[dev][acpi] add acpi bus driver
Step 1 of ACPI bus driver. Define ENABLE_ACPI_BUS=true in
a build option to use the ACPI bus driver instead of the
acpisvc service. The ACPI bus device is the root of the
system that uses ACPI. It initializes the kernel PCI driver
and publishes the PCI root device (formerly responsibility
of devcoordinator).

Next step is to publish more devices from ACPI.

Change-Id: I981c096fab8d45c74217f75a14102344a5ad6500