Use 0 as the resource id for import delegates.
Since import delegates are not present in the resource map, it doesn't
really matter what id we give them so we might as well give them a
sentinel value which makes the debug output easier to interpret.

Zero is a better sentinel than max-int anyhow since clients are forbidden
from creating resources with id 0 but they are allowed to create resources
with id max-int.

Change-Id: Ie177121527a216b767e709ec5a2cc8db99b2512a