[dev][platform-bus] Add support for mapping MMIO regions and IRQs
MMIO regions and IRQs can now be specified in the MDI for platform devices.
The mmio and irq nodes in the MDI are translated into resource handles,
which are later passed to mx_vmo_create_physcial and mx_interrupt_create.
For now this is not really necessary since the platform devices run in the same
devhost as the platform bus driver, but will be needed later when we run
platform devices in separate processes.

Change-Id: I7f088b77257737c05a945de20b52e0d74ddce8b2