Fix the build.
A recent sync-up in third_part/skia changed the method signature of
onGenerateTexture.  Update the override here in order to fix the
build.

Note: someone should probably come through here and make sure that
this implementation is doing the right thing with the addition of the
new parameter.  This CL fixes the build, but ignores the new param.

Change-Id: I6ca1b6dd892c264eb680d3eef30861b01f94ba3a