[syscalls][jobs] Add mx_job_set_relative_importance
Updates a partial ordering between jobs so that job A will be killed before
job B in low-resource situations. Will be used by the out-of-memory (OOM)
ranking service to tell the kernel which jobs to kill first.

No actual OOM killing happens yet.

Also adds the 'k mx jobs' command which dumps the list of all jobs in
importance order.

Change-Id: I818dd4e6bdb891a6cc4b4468215d23d479fa58f6