[dev][platform-bus] Start on device specific implementation support
The platform bus driver provides generic support for loading a number
of platform devices and can provide MMIO regions and IRQs to these
devices based on information in the MDI.
This change takes the first step toward expanding this to support
platform specific services like GPIOs, clocks, I2C etc.

To do this, the platform bus driver will require a platform specific
driver to bind against it. This driver will make these additional services
available to the platform bus driver, which will then make them available
to platform devices via the platform device protocol.

In this change we are mostly just adding the skeleton for this new structure.
APIs for GPIOs, I2Cs, etc. will be added in subsequent changes.

The platform specific implementation driver binds to the platform bus driver
by matching against the MX_PROTOCOL_PLATFORM_BUS protocol and the
BIND_PLATFORM_DEV_VID property.
In this change, the bcm-mailbox driver now binds as the platorm specific
bus driver rather than as a platform device.
In subsequent changes, the BCM GPIO and I2C suppport to the mailbox driver.

As soon as the platform specific driver binds, it registers a pbus_interface_t
protocol with the platform bus driver, which will be used for providing
all its platform specific services to platform devices.
The platform bus driver will wait until the pbus_interface_t is registered
before adding the platform devices to ensure that all necessary services
are available when drivers bind to the platform devices.

The platform bus driver can also provide MMIO and IRQ resources to the
platform specific driver. This is done by specifying them in a new "bus" node
in the MDI.

This change also refactors the platform bus driver into multiple source files
and replaces the platform device register_protocol/find_protocol mechanism
with a get_protocol API in the pbus_interface_t protocol.
The bcm-mailbox driver now uses this for providing the bcm bus protocol
to the BCM platform drivers.

Change-Id: Ibfe5caeae425d9d97edd20760d3d7724e8d4f1bb