apps/media: rename framework 'part' to 'node'
In the in-proc streaming framework used in apps/media, a source,
sink or transform in a graph was referred to as a 'part'. While
there aren't great alternatives to this term, 'part' is particularly
awkward. This CL renames 'part' to 'node'. While 'node' isn't
descriptive, it does fit with the term 'graph'.

Other alternatives consider (for the curious): filter, operator,
processor.

Change-Id: I9f54d35dd01da85a01af2e3088f4c055c35fa90b