apps/media: input/output references in the framework
The in-proc streaming framework used a scheme for referencing inputs
and outputs using a pointer to the Stage that owns the input/output
and an index. Input and Output classes didn't have a pointer to the
Stage that owned them and didn't know their own indices. The
motivation for this seemed reasonable at the time, but it really
doesn't make any sense.

With this CL:
1) An Input or Output knows its own index.
2) An Input or Output has a pointer to its owner Stage.
3) InputRef and OutputRef (opaque references to Input/Output for
   the API) contain Input/Output pointers rather than stage/index.
4) Input/Output pointers replace InputRef/OutputRef for all internal
   references.

Change-Id: I52f7105746c8b11fb3d5e646bae36fd2f2e040d7