Fix semaphore handling in pause/resume
When we pause page flipping, multiple page flips are deferred
and semaphores (both wait and signal) are reused amongst them.
So on resume we should handle each of the flips in order (wait,
then signal); signalling the earlier semaphores will unblock
rendering that is in the pipe, the completion of which will
signal semaphores gating page flips further down the list.

MA-215 #done

Change-Id: I42059c2c5ea1fb26fa37d1de5290f9b631a75ea4