Always require a TokenProvider to enable sync.
This CL changes the behavior of GetRepository on
LedgerRepositoryFactory. Now, it is mandatory to provide a TokenProvider
service to enable sync. As benchmarks use sync but without
authentication, this CL also provides a fake implementation of a
TokenProvider suitable for unauthenticated Firebase instances.

Now, GetRepository and EraseRepository have the same behavior regarding
TokenProviders.

LE-260 #done

Change-Id: I94026c9a82f36bed06abde18262f933e12bfb4e3