[syscalls] mx_handle_close(MX_HANDLE_INVALID) now returns MX_OK
In the face of MX_POL_BAD_HANDLE, a bad handle error is fatal, but
closing the "known never a handle" handle is similar to free(NULL)
and is a reasonable strategy for cleanup.  It's a different situation
than "program double-closed a handle" or "program closed some random
not a handle thing".

Change-Id: I25410ee3040ec062700a4ca4d363841ca2b623c0