Add jiri_update_if_green.py script
The jiri_update_if_green.py will check if the fuchsia build is green
and then perform jiri update. It will show an error message when the
build is red and not perform the jiri update.

The default builder to check is set to "fuchsia_x86-64-linux-debug".
If //scripts/env.sh is used and fset() is called in the current shell,
the current build variant is used to determine which builder to check.

Change-Id: I189e6faa47ef9da543f4839a295c087dde7e0483