[dev][platform-bus] Replace MDI nodes for MMIOs and IRQs with a list and array
Platform bus and device nodes used to contain individual nodes for MMIOs and IRQs.
Now we have an "mmios" list containing all the MMIO resources
and an "irqs" array of IRQ numbers.
This makes the index of the MMIOs and IRQs in the MDI more explicit than having
individual "mmio" and "irq" nodes interspersed within the "bus" or "device" nodes.

Change-Id: I6323e94788f046c653b73968286f6baa9b360700