Modesetting: Implement enough for bringing up a DisplayPort display
Program enough of the registers that we can display something on some
displays.  This calculates some of the parameters (primarily CRT
timings) based on the EDID data, so it will work on displays of
different sizes.  Some other parameters are hard-coded for the time
being.

Move the example EDID data from test_edid.cc into a separate file
(example_edid.{cc.h}) so that it can be reused by test_displayport.cc.

MA-150

Change-Id: Ib9cc8339a477f158c2645eeb6b6ec3b88c3d479e