Separate logging functionality out to separate library
This splits FTL's log functionality out of FTL into a separate
library that FTL depends on so clients that want to use logging (such
as fidl-generated bindings code) but not the rest of FTL can express
their dependency more tightly. This is mostly a matter of build changes
but there is one client-visible change - the command line parsing
helpers for setting log settings are moved from
//lib/ftl/log_settings.h to //lib/ftl/log_settings_command_line.h to
allow clients who aren't interested in command line parsing to still
have and set a log level.

Change-Id: Idcea3867795b2bc50eefd3253e25105b4e2e1dbe