Implement elementary cloud dashboard for Ledger sync.
This patch implements a simple cloud dashboard for Ledger, allowing to:

 - authenticate
 - see how many devices are synced via Ledger
 - erase / reset the Ledger state

This is proposed as the new way to erase Ledger, to replace existing
integrations in device shell.

Any devices that are syncing while the cloud is erased detect the erase,
delete local data and sign the user out.

Any devices that are not signed in will detect the erase upon the next
sync attempt, delete local data and sign the user out.

In any case Ledger is usable again from the next sign in.

LE-265 #comment add an elementary cloud dashboard

Change-Id: I311a6dd6484d63deb7435b27a9a160635c9f9aaf