Bring back the counter example test.
This was part of the user shell test but was deleted
from it in https://fuchsia-review.googlesource.com/#/c/39197/.

Add the deleted part of user shell test as its own test
`link_data', and delete the dead code left behind in the
user_shell test.

Add a few test points to verify that the counter really
counts. There should be more testing of link behavior in
the future.

FW-234 #done

Change-Id: Id394d963a29f66e7bbe70081a5b7a5ea89449bed