[jemalloc] Teach jemalloc to allocate out of a single VMO
This patch teaches jemalloc to allocate out of a single VMO the same
size as its pages VMAR.

jemalloc maps new pages in two ways.

In the first, it wants to extend an existing mapping. In this case we
can easily compute the offset into the VMO, as it is the same as the
offset into the VMAR.

In the second, we want a random placement from the VMAR. To do this,
we create a subvmar of the size we want, destroy it, and map the VMO
over the same spot. In the future, we may tweak this to allow mapping
the VMO directly without the intervening call to _mx_vmar_destroy of
the subreservation. This would allow us to remove the userspace
lock. In the meantime, it's not an especially expensive lock to take,
as modifications to the VMAR are going to be serialized by the kernel
anyway. MG-942 tracks the suggestion to modify MAP_SPECIFIC regarding
this behavior.

This patch will also allow us to take advantage of adjacent VMO
mapping coalescing in the future, cf MG-941.

This patch also teaches jemalloc that we are overcommit, which allows
simplifying the map code paths.

Building all of Fuchsia and doing things until we run out of memory
for framebuffers worked with no malloc issues.

US-302 #done
MG-907 #done

Change-Id: Id85150470be663e8b06b976ccb04b0b9cbfd1f16