[futex] Combine RemoveFromHead()+WakeThreads() into a single operation
Before this change, RemoveFromHead() detaches a list of threads to be
woken, and WakeThreads() goes through the list to wake those threads.

Combine that into a single operation: Add a WakeThreads() function
that is similar to RemoveFromHead() except that it wakes threads
instead of constructing a list of removed threads.

This might be more efficient because it involves only one pass across
the list instead of two, and it involves less pointer manipulation.

Note that this removes the oddity from the old WakeThreads() of the
comparison "node != head" where head could be a dangling pointer.
(Instead, we save the result "is_last_node = (node == list_end)"
before node could be freed.)

Change-Id: I0d986e646f44fd0037247fe8fff6bcb2b0807c58