[block] Prevent fifo from attempting to read from closed handle
Previously, this code actually relied on a mechanism where a
server thread (periodically reading from a fifo) could be
terminated by another thread by shutting down that fifo. On the
server's next attempt to read, the fifo read would (predictably)
fail, and the server would terminate.

With the addition of MX_POL_BAD_HANDLE, this behavior is more
catastrophic: when attempting to access an invalid handle, the
entire server process is terminated.

This patch fixes the behavior by introducing a "terminate" signal
which is used as an explicit message to ther server, rather than
"closing the fifo". The server's end of the fifo is kept valid
until the BlockServer destructor executes.

MG-944 #done

Change-Id: Icdb373dfeb6ae56bed57b9ff499849b0f1909a0e