[magenta] introduce MX_RIGHT_SIGNAL and MX_RIGHT_SIGNAL_PEER
Previously, the ability to set/clear the user signal
bits on an object via mx_object_signal() was controlled
by MX_RIGHT_WRITE and the ability to set/clear the user
signal bits on the peer of a channel, socket, or eventpair
was not restricted at all.

The new rights MX_RIGHT_SIGNAL and MX_RIGHT_SIGNAL_PEER
are now used to provide access control for these operations.

Change-Id: I3d82457c39c062f181615fa40266ac72eae2b802