apps/media: don't start audio_policy_service
audio_policy_service was started at boot time so that system audio
gain would be set properly before the first audio renderer was
created. Now we're making sure its started when we create an audio
renderer, so this is no longer needed. In any case, starting the
service with apps.config caused two instances to be started, because
the apps.config instance isn't recognized as a viable handler for
connect requests.

Change-Id: I8e57c3f70897062ea9033ebab52bf5be932f50dc