Update sysui to package template, split out tests
This updates the sysui target to use the GN package template which is
easier to manipulate and less error-prone (since GN will check for the
existence and dependencies of the files). Also splits sysui_tests out
to a separate package, which is currently empty as all tests are
host-only, but will likely become non-empty at some point down the line.

Change-Id: I0600c25e31c40c57b833a2087c75d618ea006899