[kernel][channels] Prevent mx_channel_call_finish() with mismatched channel
The kernel provides a pair of syscalls for implementing
mx_channel_call():
 * mx_channel_call_noretry()
 * mx_channel_call_finish()

If mx_channel_call_noretry() is interrupted by
mx_task_suspend()+resume(), userland was supposed to call
mx_channel_call_finish() with the same channel handle.  However, the
mx_channel_call_finish() syscall didn't check that userland passed the
same channel handle, it just assumed that.

Before this change, if userland passes in a different channel handle,
there are parts of the kernel code that become thread-unsafe.  The
kernel can do waiters_.erase(*waiter) when waiter isn't a member of
waiters_, and it can do this while holding the lock for the wrong
instance of ChannelDispatcher.

Fix this by removing the channel argument from
mx_channel_call_finish() so that it's not possible to pass a
mismatched arg.  Instead, we get the ChannelDispatcher reference from
the thread's MessageWaiter object.

 * Change ResumeInterruptedCall() to take a MessageWaiter so that it
   doesn't have to re-get the MessageWaiter via
   UserThread::GetCurrent().

MG-948 #done

Change-Id: Iee770e81002ddca3a7caa62998c0da75e080b9dd