Add StoryScope to maxwell::ComponentScope.
This is an initial stab at adding Story Scope to the ContextService, so
that modules can report their text through Flutter to the ContextEngine.

These reports should be scoped both by Story and by Module. This CL
takes a first stab by creating StoryScope and using it in
StoryControllerImpl. A follow-up CL to ContextPublisherImpl should use
this to prefix topics appropriately.

Change-Id: I4e65d18c1f4def4889087c7eb9c9dbf0cff9678a