[ethernet] Fix to avoid closing a FIFO handle while it is in use
This avoids a race that can cause a use-after-close of a handle, which
could cause an MX_POL_BAD_HANDLE exception once that policy is enabled
for drivers.

The issue was that eth_tx_thread() calls mx_fifo_read() on the handle
edev->tx_fifo, but this handle can also be concurrently
mx_handle_close()'d by eth_kill_locked().  This creates a race
condition: This can cause an mx_fifo_read() of a closed handle or of
MX_HANDLE_INVALID.

This fix uses the same approach as the fix for MG-944.

I tested this by doing:
  run-magenta -n ... # Run Magenta under QEMU with a network device
  netreflector &
  ps
  kill <pid of netreflector>
  # The driver should print "eth [netreflector]: tx_thread: exit: 0"

MG-946 #done

Change-Id: Idcd6af4be516c833ab766ae8bd7b87aa680be016