[ddk] Clean up driver ELF note definition
The ELF note header's n_namesz field includes the '\0' terminator
of the name, so 7 was wrong for "Magenta".  Clean up the initializers
of the namesz and descsz fields to use entirely sizeof-based arithmetic
with no hand-written constants.  Split the header into substructures
to make the layout constraints clearer.

Also, make sure the ASan build doesn't introduce extra padding into the
note section, which violates the ELF format rules for notes.

Change-Id: I9ad380dafe1e8261f10e3ff41ca3d0c9769e0604