[runtime] Teach packages to use runners.
This allows a .dartx or .flx file to be placed in a Fuchsia package
at `bin/app`. This will allow launching the contained Dart/Flutter app
when the package is launched.

It does *not* do sandboxing, or allow the Dart app to access files
within the package, other than those already contained within the .dartx
or .flx archive. There are other bugs (US-312, US-313) filed to track
progress on those features.

US-307 #done

Change-Id: I6237b463d7350b35ca105edacbd0cbbf4bd3bb58