[ulib][mxio] New loader-service operation LOADER_SVC_OP_CONFIG
This new RPC lets a process tell the loader service about its
configuration that might be relevant to library searching.

For now, the mxio loader service treats the configuration
string as a subdirectory under ../lib/ to search in first.
If the string ends with '!', look only in that subdirectory (sans '!').

MG-71 # Loader service protocol to support sanitized shared libraries.

Change-Id: I6e8e4e0aca816af91076467ff7bcb8ed749486dc