app: Keep repo handles alive while repository is being erased
This patch ensures that while repo state is being erased, the repo
itself is shutdown, but:

 a) handles bound to it are kept alive
 b) any new requests to bind to this repo are failed

This ensures that no operation can be performed on the repository while
it is being erased.

The patch also fixes erase recovery apptest and adds one for the
EraseRepository method.

LE-264 #comment keep repo handles alive while repository is being erased

Change-Id: I070ad205bd795738c272f608f81738677f4b9336