[build] Support AddressSanitizer userland build with USE_ASAN=true
ASan builds by default go in build-$PROJECT-asan/.
The buildall and run-magenta scripts do ASan builds with the -A switch.

Code that is incompatible with sanitizer builds adds
$(NO_SANITIZERS) to MODULE_COMPILEFLAGS.  Everything else
in userland gets built with ASan.

Using USE_ASAN=true requires a Clang toolchain with the ASan
runtime shared library, which will not land for a while yet.

MG-71 # Enabled all-or-nothing USE_ASAN build switch for Magenta standalone.

Change-Id: Ifc0628934c19affcb909b20f3cfdcc86ee54c2f9