[mini-process] Move the subprocess's code into a separate ELF object
Before this change, the code for minipr_thread_loop() was copied from
the parent process into the child process.  That was a hack that
assumed that the function was no bigger than kSizeLimit bytes.  If
minipr_thread_loop() contained any calls to external functions (maybe
inserted by the compiler), that could cause crashes at run time.

Clean this up by moving that code to a separate ELF file (an
executable/library) and by ELF-loading that.  This means that any
problems of having external references will be caught at link time.
This also makes it possible to split up the subprocess code into
multiple functions.

Change-Id: I1ecd19ae5457c3753080fd1695e0acde2d5af08b