[unittest] Adds ASSERT_DEATH, for declaring a function to crash. This is so users don't need to create a separate process.
Instead of the crash handler being in a new thread, the test function is
run in a new thread so we can kill it if it crashes (and the rest of the
tests keep running).

When the user calls ASSERT_DEATH on a function, we create a new thread
and bind it's exception port and register it as expected to crash,
before running the function.

The original RUN_TEST_ENABLE_CRASH_HANDLER / REGISTER_CRASH
functionality will still work too.

MG-802

Change-Id: Ia91b548b376a439d0ed06c4b3dbb873112c76518