[libfs] Plumb fs::Dispatcher through filesystem servers.
Replaced vfs_rpc_serve with Vfs::ServeFilesystem which takes an
fs::Dispatcher.  As part of this change, we also make fs::MxioDispatcher
a little more flexible about which thread it runs on.

This will be followed by a change which plumb the calling dispatcher
through as an argument to Serve() instead of requiring each Vnode to
carry it (unless it wants to serve somewhere else).

Change-Id: Iac53e00d8eaf9d1bd781b2f9b56debc87a49fb8b