[build] Support ENABLE_ULIB_ONLY=true make option
In this mode, the build produces nothing but a bootdata.bin that
places the shared library binaries into /boot.  This will be useful
in a USE_ASAN=true build to provide the lib/asan/ libraries needed
by ASan builds of other components.

Since it doesn't build the host tools necessary to build some of the
libraries, it requires giving TOOLS=... on the command line to point
at the build-.../tools subdirectory of some other build.

Change-Id: Ie6dca88a1f55ca2839cf14797d6d876a07c4ac4b