[kernel][vm] Fix edge-case in range intersection checks
This fixes a race hit in MG-949, in which one process performs an
mx_vmar_unmap() call to completely unmap a VMO mapping while another
process calls mx_vmo_op_range(COMMIT) on that VMO.  During the unmapping
process, the size_ of the VmMapping in the first process gets set to 0
right before the VmMapping is destroyed.  It briefly drops the VMO's
lock before removing itself from the VMO's list of mappings.  In this
window, the mx_vmo_op_range(COMMIT) can cause UnmapVmoRangeLocked() to
get called on the VmMapping.  This uses GetIntersect() to check for
intersection, which goes awry when size_ = 0.

MG-949 #done

Change-Id: I0f7a13e7d37b4296c382dcc6475b9386ff2a014b