Take advantage of zx_port_cancel.
Take advantage of zx_port_cancel and the guarantees it provides to
transfer managed references to DispatcherEventSources to wait_async
jobs which have been posted to ports.  This eliminates the need to
generate unique tokens (which had been used as the key for wait_async
jobs) and then map from token to event source every time a job
completed.

Change-Id: Ic484dfe1efebacf21d42d0bd0bf707e0651e2ed9