apps/media: implement free-threaded update request scheme
In the in-proc streaming framework, stages could be added to one of
two update backlogs in two different ways: 1) aynchronous to a
Stage::Update call, Engine::PushToDemandBacklog or
Engine::PushToSupplyBacklog could be called, or 2) on on arbitrary
thread, Stage::RequestUpdate could be called. The first method was
safe, because only one stage could be in Update at any given time.

This CL implements an alternative scheme in which there's a single,
thread-safe way to add stages to the backlog. A few other related
changes have been made as well.

There's only one backlog now that works in FIFO order. This may not
be the most efficient approach, but we'll change that later if need
be.

To add a stage to the update backlog, Stage::NeedsUpdate may be
called on any thread. NeedsUpdate uses an atomic counter to determine
when the stage actually needs to be enqueued (i.e., it's not already
enqueued). In addition, the scheme detects the needs to re-run Update
due to changes made concurrently with Update.

The Input class has also been changed to require the downstream stage
to explicitly 'take' a packet rather than just std::move'ing it. This
lets the Input know that the upstream stages 'NeedsUpdate'. This
replaces a hacky scheme previously implemented in Engine.

This CL also introduces a dispatching scheme that will eventually
allow the graph/engine to be driven by e.g. a port that dispatches
multiple threads. This involves an update callback out of the engine
indicating the need to do updates. For now, the graph is just calling
back into engine on this callback so we're still single-threaded.

Change-Id: I70a97623406c5657a6d03b65977c365c006f7206