[ulib][musl][ldso] Fix dlopen and dlsym with indirect DT_NEEDED dependencies
Symbol resolution needs to follow the whole dependency graph, not just
one level deep.  This matters for dlopen to resolve references to
indirect dependencies at load time.  It also matters for dlsym to find
the named symbol anywhere in the dependency graph of the given object,
not just in the object itself or a direct dependency.

MG-963 # done

Change-Id: I9770e81937050248ed73d1185311246cc9b4ac60