[hypervisor] Make linux serial output work more.
Linux kernel serial output is done with busy-waiting on LSR.
Userspace serial output prefers writing to buffers and handling THR
empty interrupts.

This change basically adds support for firing THR empty interrupts.

Things to consider:
1) We currently emulate a simple 8250, with no buffer whatsoever
(because it's easier than jumping through the hoops to appear fancier).
This causes an interrupt to be fired for each character, because of
this...
2) We cheat a little on interrupt_id, leaving it set when it should be
cleared so that we don't need to take a lock in the output loop to set
it. This is technically wrong, but it's probably not much of an issue
since...
3) Once we get virtio-console support all of these problems will go away
and...
4) virtio-console is what I'm working on next.

Change-Id: I5f645bbdfebfdc881357aab5890d94bf5795408c