Use report_result to run cobalt_apptests
Our test harness has the ability to use arbitrary
binaries as tests as long as the binary returns 0 for success and
non-zero for failure. One invokes the binary via the |report_result|
binary. In this CL we

- change cobalt_client_tests.json to invoke cobalt_apptests via
report_result rather than directly.

- We change cobalt_apptests so that it returns 1 on failure.

Also I removed the code that uses the MessageLoop since all FIDL
calls are synchronous. However it appears it is necessary to
*have* a MessageLoop created on the main thread.

Change-Id: I27f406bb00fd6c52dd9ddf78443d8924962906d4