apps/media: switch to shared PacketPtr
This CL redefines PacketPtr as a shared_ptr rather than a unique_ptr.
This will be needed for 'distributors', that supply the same packet
to multiple recipients. It also allows us to leverage the template
specializations for atomic operations.

PacketPtr previously had a custom deleter that called the virtual
method Packet::Release. This was originally intended to allow the
use of a lookaside, but that never happened. We still need code in
the destructor to dispose of payloads.

std::make_shared is used throughout to eliminate the double
allocations otherwise associated with shared_ptr. Packet constructors
are now public as a consequence.

Change-Id: Id10010aca5addc32d8355adbdd1a39cf3a96a519