Introduce the first fun service: Sketchy.
Sketchy starts from the Escher Sketchy Demo and massages it
to work well with the Mozart SceneManager.  In particular,
the "kWobble" shape-modifier is not something that we want
to expose via the Session API; it is not generally useful.

Instead, the Sketchy service will use the GPU to generate
new meshes each frame.  The result will be indistinguishable.

Change-Id: I01afcada09d63429fc170055d44b7d658b5f668f