[fs] Improve vfs_lock_ encapsulation
- Convert thread annotations to no-op on builds with
"DISABLE_THREAD_ANNOTATIONS" defined; this allows static thread
analysis for operations which act on filesystems that
are compiled for both host and target (such as MinFS), but
which use a single-threaded model for host-side tools.
- Add thread annotations to VFS operations
- Continue refactoring required to eventually make "vfs_lock_"
internal to the VFS class. "Mount + Mkdir", "Locked" functions,
and iostate token access are now all internal to the Vfs class.

Change-Id: I5d083f38726c493602cd24139303ab377ef655e4