[hypervisor][pci] Fix reading PCI BAR size on linux.
The IO/MEM enable bits in the command register control if a device will
respond to IO access. Linux will disable IO accesses before measuring
the BAR size to ensure that the PCI device doesn't respond to any
requests while mapped to the temporary value used to measure the BAR.

This means we do want to make sure we still correctly mask out the
lower bits in the BAR to allow software to measure the BAR and also
ensure we don't forward any IO requests to the device when the
corresponding command register bits are clear.

Change-Id: I12a374a21f369018747d9ffe25973a7b54d106f6