apps/media: change framework stage interaction for multiproc
This CL changes the way packets and demand are communicated between
framework inputs and outputs. The goal is to allow adjacent stages
to be updated concurrently. The new scheme uses an atomic value in
Input that encodes both the demand and the fullness of the packet
mailbox (also in Input).

Change-Id: Ia9e40d5ed28a1f6ebead0da538abe8c0d5c48f9d