Convert LinkImpl to OperationQueue.
All FIDL operations of LinkImpl are serialized, including SetSchema() and
Watch() / WatchAll(), such that they are executed in call order wrt Set(),
UpdateObject(), and Erase(). Thus, Set() followed by SetSchema() will apply
the schema to the value that was just Set(). Set() followed by Watch() will
yield an initial Notify() with the value that was just set. Etc.

Notifications reveiced from ledger are ignored while a write operation is
pending, which addresses FW-208. This may lead to network updates being
ignored if they are received while a write operation is in progress. This
doesn't change the situation compared to status quo that updates on different
devices may clobber each other, and this needs to be addressed separately,
by using a CRDT or the like internal representation of the link value and/or
using a conflict resolver for link values.

To ensure that the ledger update that corresponds to the write operation
is received (and thus ignored) while the write operation is pending, we
employ the ledger hack to start an empty transation after the write. A
StartTransaction() call only returns after all pending watcher notifications
are delivered.

Enable the test that was broken by FW-208.

FW-208 #done
FW-251 #done

LE-278 #comment This change here can be improved once that issue is addressed.

TODO in followups:

* Code is somewhat duplicated between Operations (as it was between methods
  before).

* It's not clear how to test effects of remote ledger updates. We could have
  a way to inject such changes directly through StoryProvider to make tests
  very easy.

* Conflicts need to be resolved by some form of merge rather than clobbering
  the whole link value.

* We need more tests of Link operations. The newly added test can be extended
  to cover more Link bahavior.

Change-Id: Idc9311d616837e793c32af49cbd2375514f801bd