Fix link connection id for self notification suppression.
Instead of a weak ptr, use a monotonic counter value to identify
the LinkConnection instances of a LinkImpl. An ID is never reused,
which was the situation the weak ptr was meant to guard against.

This solution is more correct, because previously the source was
identified by a naked pointer, which could potentially be reused
between the storage write starting and the notification being sent.

Also, nullptr was used as source connection for changes received from
the ledger, notifications of which would thus have been suppressed
on WatchAll() watchers and all watchers whose connection was closed.
Using a special ID to identify the source of ledger changes solves
this problem as well.

Change-Id: I877c286d1e6e1aa682dd7cc1dadb88096e2d7cef