Add git-file-tidy script to run clang-tidy on all modified files.
The tool is like the existing `git-file-format`, but for linting using
clang-tidy. It takes care of generating the compilation database and
only lints the files that are modified, staged, or were touched by
commits not in the tracked branch (local commits).

This patch also changes/fixes the behavior of git-file-format --all, so
that only files in the subtree of the current working directory are
formatted.

LE-277 add git-file-tidy

Change-Id: Ie5fa9bee2f1f11e706704fc17c365e16e581cd6b