StoryProvider.GetLinkPeer().
GetLinkPeer() obtains a connection to a link that is direct to
the ledger, bypassing the StoryControllerImpl, StoryStorageImpl,
and LinkImpl instances already existing on the current device.

Updates made on such a Link connection behave when seen from Links
used in a running story as if they come from another device.

This makes it possible to write tests for remote updates of Links.
However, it's still not possible to test the behavior recently
implemented in
https://fuchsia-review.googlesource.com/#/c/modular/+/44909/, as
there is still no way to force a remote update to arrive at a certain
time, and to collide with a local update.

Extended the existing link_update test by what's possible now,
leaving notes as to what's still not covered.

This is EXPERIMENTAL with regard to the practice to build features into
core code that are there only to support development and testing. I'm
willing to do this for now because I have a hunch (1) that link peers
might be useful for other things than testing, and (2) that it's rather
important to have ways to easily simulate remote updates and their
reconciliation especially in light of the upcoming CRDT work by jimbe@
on Link. In the long term I expect this to evolve to look differently.

Change-Id: Id698a10b7cddd09a46d5f1a90fbc977842598823