[boringssl] Update GN files.
This CL introduces new GN files to match the post-FIPS structure of
//third_party/boringssl.  The python script was becoming increasingly
brittle, so it has been retired for the time being.  This means the
BUILD.gn files will need to be manually updated until the cmake GN
generator is ready.  To make this easier, the GN files here match the
layout and organization of the CMakeLists.txt files within BoringSSL.
Additionally, this CL includes a trampoline script for running go and
perl scripts.

Change-Id: I1ec193696066e4edd5d59bea459d246b7cd3ef54