Initial FIDL interface integration
Added Scan interface and a command tool to call it.
The Scan interface is still preliminary and likely to change.

As before, if the config file specifies a SSID, wlanstack will
scan it repeatedly and try to join it. During the scan interval,
a FIDL Scan command can be accepted and run. Once the association
with a SSID is successful, a FIDL Scan command will start returning
an error message that Scan in the associated state is not currently
supported.

If the config file doesn't specify a SSID, no repeated scan is
started. A FIDL Scan command can be accepted and run.

Change-Id: I7bbca59eeb359d22e5057baa7e9c22f024e67081