[entity] v0 Entity Model interfaces.
This creates:

* EntityFactory
* EntityResolver
* Entity
* EntityReference
* EntityContentType

and begins to describe the relationship between them.

EntityResolver and EntityFactory are meant to be exposed to processes
through their namespace.

EntityResolver should eventually route its
GetEntity() calls to an appropriate EntityProvider (which is not
included in this CL).

EntityFactory, for now, will also be exposed through the namespace,
although it should be replaced with other mechanism for creating
Entities, and the interface as is now should be used for creating
"transient" or temporary Entities.

FW-257 #done

Change-Id: Ic3bf4868d9c351cb255c1fb3eddcc2df6a3da4f7