[unittest] Suppress warnings from static analyzer when assertion failed
This patch added a function 'unittest_returns_early()' and it
will be called when an assertion failure occurrs. The
MagentaHandleChecker from static analyzer will treat calls to this
function as sink holes and suppress warnings on handle leaks bugs.
It will help reduce the false positives when static analyzer is
run on unit tests as most tests do not clean acquired resources when
assertion failures occur.

Change-Id: I09c13c428a76400b4e2c86d67a893a36d0b7dd64