[kernel][vm] Implement zero-permission mappings
We were seeing if we could avoid having a PROT_NONE equivalent in favor
of using VMARs to provide guard pages, but we have found that the VMAR
solution does not address some less common use patterns.  In particular,
the use case of marking pages as inaccessible and later marking them
accessible was not addressed.  This behavior could be implemented using
the existing system, but would require additional error-prone bookkeeping
in each usermode app that wants this behavior, and would preclude some
optimizations that could be done in the kernel around coalescing
adjacent mappings.

MG-426 #done

Change-Id: I94ff7b62bb516103ff5a27623d7039697868c463