[kernel][timer dispatcher] Fix DPC race condition a better way.
With this change, we enforce the invariant that we only reset the timer
when there isn't already a timer callback pending.  So we always cancel
the timer before resetting it, and if cancelation fails (because the
timer callback is pending) we defer the reset until that callback runs.

This change fixes the DPC races which were previously worked around as
well as a new one which hadn't been addressed.  That race could happen
if TimerDispatcher::Set snuck in while TimerDispatcher::Cancel was waiting
for the callback to run.  The window of opportunity for this was quite
small and we have not observed the problem but it was possible.
And now it isn't.  :)

Change-Id: I62dcdf2e9bfc465b16bcc161e2d3004273967181