[kernel][timer dispatcher] Optimize timer signalling.
If the timer is already due by the time TimerDispatcher::Set runs, set
the timer signal immediately instead of scheduling the timer callback.

This optimization benefits message loops which use timers to keep track
of the time when their most recent task is due.  With this change, they
are guaranteed not to block during the next iteration of their wait loop.
Previously they might block (briefly) if the timer callback had not yet
fired.

Also ported the timers tests to use mx::timer.

Change-Id: I013f65e177639b1b3d4d14e81be20e7dedf7b1c2