First unittest for AgentRunner.
- This tests connecting to an agent will start a new application if no
connections exist already, and that connecting to an existing agent
won't try to start a new one, or try to re-initialize the agent.

- This test uses a FakeApplicationLauncher, which instead of launching a
new process, allows you to consume the ApplicationLaunchInfo and
ApplicationController and drive them yourself.

- LedgerRepositoryForTesting starts a Ledger process and acquires a
LedgerRepository. Everything else in this test happens in a single
thread, on one message loop.

Change-Id: I8b844b18e38f5b6e44dc3b0a72667770430b1cbf